review after session - start refactoring
