+++ assets/js/cart.js
@@ -1,0 +1,14 @@
+import * as index from "./index";
+
+////////////////////////////////////////////////
+//Button decrement and increment with my input//
+////////////////////////////////////////////////
+index.incrementDecrement();
+////////////////////////////////////////////////
+//Button decrement and increment with my input//
+////////////////////////////////////////////////
+
+
+console(index.getCart()); //je récupère mon panier
+
+function renderCart()
